move fetchdata to app folder
